add validation

+++ src/assessment/assessment.controller.ts
@@ -6,11 +6,14 @@ import {
   Param,
   Patch,
   Post,
+  UnauthorizedException,
   ValidationPipe,
 } from '@nestjs/common';
 import { assessmentService } from './assessment.service';
 import { AssessmentDTO } from './dto/assessment.dto';
 import { UpdateAssessmentDTO } from './dto/update-assessment.dto';
+import { CurrentUser } from 'src/auth/decorator/CurrentUser.decorator';
+import { UserPayload } from 'src/auth/types/UserPayload';
 
 //o controlador lida com as requisições http
 @Controller('assessment')
@@ -18,7 +21,10 @@ export class AssessmentController {
   constructor(private readonly assessmentService: assessmentService) {}
 
   @Post()
-  async create(@Body(ValidationPipe) assessmentDTO: AssessmentDTO) {
+  async create(@Body(ValidationPipe) assessmentDTO: AssessmentDTO, @CurrentUser() currentUser: UserPayload) {
+    if (assessmentDTO.userId !== currentUser.sub){
+      throw new UnauthorizedException('Só é possível criar posts para si mesmo');
+    } 
     return this.assessmentService.create(assessmentDTO);
   }
 
@@ -34,14 +40,23 @@ export class AssessmentController {
 
   @Patch(':id')
   async update(
-    @Param('id') id: string,
+    @Param('id') id: number,
     @Body(ValidationPipe) updateAssessmentDTO: UpdateAssessmentDTO,
+    @CurrentUser() currentUser: UserPayload
   ) {
+    const assessment = await this.assessmentService.findOne(id);
+    if (assessment.userId !== currentUser.sub){
+      throw new UnauthorizedException('Você só pode atualizar seus próprios posts');
+    }
     return this.assessmentService.update(+id, updateAssessmentDTO);
   }
 
   @Delete(':id')
-  async remove(@Param('id') id: string) {
+  async remove(@Param('id') id: number, @CurrentUser() currentUser: UserPayload) {
+    const assessment = await this.assessmentService.findOne(id);
+    if(assessment.userId !== currentUser.sub){
+      throw new UnauthorizedException('Você só pode excluir seu próprio post');
+    }
     return this.assessmentService.remove(+id);
   }
 }
